move language support package to modern ts
