+++ packages/language-server/package.json
@@ -50,6 +50,7 @@
   },
   "devDependencies": {
     "@types/lodash.debounce": "^4.0.9",
-    "esbuild": "^0.25.1"
+    "esbuild": "^0.25.1",
+    "typescript": "^5.8.2"
   }
 }
