+++ packages/language-support/package.json
@@ -48,8 +48,8 @@
   },
   "scripts": {
     "gen-parser": "cross-env ANTLR4_TOOLS_ANTLR_VERSION=4.13.2 antlr4 -Dlanguage=TypeScript -visitor src/antlr-grammar/CypherCmdLexer.g4 src/antlr-grammar/CypherCmdParser.g4 -o src/generated-parser/ -Xexact-output-dir",
-    "build": "npm run gen-parser && npm run bundle-antlr4-c3 && tsc",
-    "bundle-antlr4-c3": "cp -r ../../vendor/antlr4-c3/dist/esm vendored-dependencies/antlr4-c3",
+    "build": "npm run gen-parser && npm run bundle-antlr4-c3 && tsc --project ./tsconfig.json",
+    "bundle-antlr4-c3": "mkdir -p vendored-dependencies/antlr4-c3 && cp -r ../../vendor/antlr4-c3/dist/esm/* vendored-dependencies/antlr4-c3",
     "benchmark": "tsc --project tsconfig.benchmark.json && BENCHMARKING=true node dist/tests/benchmarks/benchmark.js",
     "clean": "rm -rf {dist,src/generated-parser,vendored-dependencies}",
     "test": "vitest run"
@@ -57,6 +57,7 @@
   "devDependencies": {
     "@types/benchmark": "^2.1.5",
     "benchmark": "^2.1.4",
+    "typescript": "^5.8.2",
     "vitest": "^2.1.9"
   }
 }
