+++ packages/schema-poller/package.json
@@ -5,13 +5,13 @@
   "author": "Neo4j Inc.",
   "license": "Apache-2.0",
   "version": "2.0.0-next.17",
-  "main": "./dist/cjs/src/index.js",
-  "module": "./dist/esm/src/index.js",
+  "type": "module",
+  "main": "./dist/src/index.js",
+  "types": "./dist/src/index.d.ts",
   "exports": {
     ".": {
-      "require": "./dist/cjs/src/index.js",
-      "import": "./dist/esm/src/index.js",
-      "default": "./dist/cjs/src/index.js"
+      "import": "./dist/src/index.js",
+      "types": "./dist/src/index.d.ts"
     }
   },
   "repository": {
@@ -31,9 +31,7 @@
     "neo4j-driver": "^5.12.0"
   },
   "scripts": {
-    "build": "concurrently \"npm:build-esm\" \"npm:build-commonjs\"",
-    "build-esm": "tsc --module esnext --outDir dist/esm",
-    "build-commonjs": "tsc --module commonjs --outDir dist/cjs",
+    "build": "tsc",
     "clean": "rm -rf dist"
   }
 }
