+++ tsconfig.base.json
@@ -4,8 +4,8 @@
     "declaration": true,
     "esModuleInterop": true,
     "lib": ["dom", "ESNext"],
-    "module": "commonjs",
-    "moduleResolution": "node",
+    "module": "ESNext",
+    "moduleResolution": "bundler",
     "skipLibCheck": true,
     "sourceMap": true,
     "target": "ESNext",
